Simpler load bookkeeping in Basic.hs
Load bookkeeping is now dealt with by functions insertIntoSlot and
deleteFromSlot exactly when mappings are deleted or added to a
hashtable.

Helper function delete' is renamed as findSafeSlots and now it only
helps find slots to which to insert / from which to delete
mappings instead of actually doing the deleting and bumping load.